requirements.txt
fastapi
uvicorn[standard]
pydantic
playwright
google-genai
flask
gunicorn
botasaurus-driver
botasaurus
botasaurus-requests
trafilatura
# Add other botasaurus deps if not included in the base image